Refactor FMData API to use PSR-4 Autoloading with separated class files
